refactor goose
