+++ crates/goose/Cargo.toml
@@ -8,6 +8,7 @@ repository.workspace = true
 description.workspace = true
 
 [dependencies]
+mcp-core = { path = "../mcp-core" }
 dirs = "5.0.1"
 anyhow = "1.0"
 thiserror = "1.0"
