+++ crates/goose/src/lib.rs
@@ -3,6 +3,7 @@ pub mod developer;
 pub mod errors;
 pub mod key_manager;
 pub mod memory;
+pub mod message;
 pub mod process_store;
 pub mod prompt_template;
 pub mod providers;
