+++ crates/goose/src/memory.rs
@@ -1,6 +1,6 @@
 use crate::errors::{AgentError, AgentResult};
-use crate::models::content::Content;
-use crate::models::tool::{Tool, ToolCall};
+use mcp_core::content::Content;
+use mcp_core::tool::{Tool, ToolCall};
 use crate::systems::System;
 use anyhow::Result as AnyhowResult;
 use async_trait::async_trait;
