+++ crates/goose/src/providers/base.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 
-use crate::models::message::Message;
-use crate::models::tool::Tool;
+use crate::message::Message;
+use mcp_core::tool::Tool;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct Usage {
