+++ crates/goose/src/providers/ollama.rs
@@ -3,8 +3,8 @@ use super::configs::OllamaProviderConfig;
 use super::utils::{
     messages_to_openai_spec, openai_response_to_message, tools_to_openai_spec, ImageFormat,
 };
-use crate::models::message::Message;
-use crate::models::tool::Tool;
+use crate::message::Message;
+use mcp_core::tool::Tool;
 use anyhow::{anyhow, Result};
 use async_trait::async_trait;
 use reqwest::Client;
