+++ crates/goose/src/providers/openai.rs
@@ -11,8 +11,8 @@ use super::utils::{
     check_openai_context_length_error, messages_to_openai_spec, openai_response_to_message,
     tools_to_openai_spec, ImageFormat,
 };
-use crate::models::message::Message;
-use crate::models::tool::Tool;
+use crate::message::Message;
+use mcp_core::tool::Tool;
 
 pub struct OpenAiProvider {
     client: Client,
