+++ crates/goose/src/systems/goose_hints.rs
@@ -4,9 +4,9 @@ use std::fs;
 
 use super::Resource;
 use crate::errors::{AgentError, AgentResult};
-use crate::models::content::Content;
-use crate::models::tool::Tool;
-use crate::models::tool::ToolCall;
+use mcp_core::content::Content;
+use mcp_core::tool::Tool;
+use mcp_core::tool::ToolCall;
 use crate::systems::System;
 
 pub struct GooseHintsSystem {
