+++ crates/goose/src/systems/system.rs
@@ -3,8 +3,8 @@ use async_trait::async_trait;
 
 use super::Resource;
 use crate::errors::AgentResult;
-use crate::models::content::Content;
-use crate::models::tool::{Tool, ToolCall};
+use mcp_core::content::Content;
+use mcp_core::tool::{Tool, ToolCall};
 
 /// Core trait that defines a system that can be operated by an AI agent
 #[async_trait]
