+++ crates/mcp-core/src/lib.rs
@@ -1,4 +1,4 @@
-mod content;
-mod handler;
-mod role;
-mod tool;
+pub mod content;
+pub mod handler;
+pub mod role;
+pub mod tool;
